Refactor delegation metrics and components to use 'sky' terminology instead of 'mkr'. Update related API calls, types, and component props for consistency across the codebase. Remove deprecated fetchV3Delegates and fetchV3DelegationMetrics files.
