+++ modules/home/components/GovernanceStats.tsx
@@ -40,7 +40,7 @@ export function GovernanceStats({ pollStats, stats, mkrOnHat, mkrInChief }: Prop
     {
       title: 'SKY Delegated',
       value: stats ? (
-        `${formatValue(BigInt(Math.floor(stats.totalMKRDelegated)), 0, 2, true, false, 1e9)} SKY`
+        `${formatValue(BigInt(Math.floor(stats.totalSkyDelegated)), 0, 2, true, false, 1e9)} SKY`
       ) : (
         <Skeleton />
       )
